Add doc to UI and visualizations components. And rename weka package as prediction and statictics package as staticticaltest
