+++ app/src/main/java/chpc/UI/views/MainUI.java
@@ -9,19 +9,26 @@ import javax.swing.*;
 import chpc.database.PostgresDb;
 import chpc.dataLoader.NHPIRecordDAO;
 import chpc.dataLoader.NHPIRecordDAOImpl;
-
+/**
+ * A singleton class can create a window when run the program
+ */
 public class MainUI extends JFrame {
   private static MainUI instance;
   private NHPIRecordDAO nhpiDAO;
   public JPanel mainContentPanel;
-
+  /**
+   *
+   * @return one window used to interact with the users
+   */
   public static MainUI getInstance() {
     if (instance == null)
       instance = new MainUI();
 
     return instance;
   }
-
+  /**
+   * private constructor used to create window containing lots of components
+   */
   private MainUI() {
     // Set window title
     super("House Price Statistics");
