+++ app/src/main/java/chpc/statisticaltest/StatisticsResults.java
@@ -1,4 +1,4 @@
-package chpc.Statistics;
+package chpc.statisticaltest;
 
 import org.apache.commons.math3.stat.descriptive.SummaryStatistics;
